clock() based timing, added '-a' arg for average
When number of elements is relatively small, CPU computes so fast that clock() cannot precisely time it. 
To address this, I first tried using std::chrono, but it turns out to be unhelpful. 
Then I realized that I could just repeat for a number of times and then take the average of it. 
In fact, this is a commonly used skill when your object to measure is too small and your measurement is too big.
And so I added a argument '-a' to the program, meaning 'average_over'.
Now the program will repeat average_over times and then use the averaged time to return.